Create pairwise-Fst-in-r
This code takes you from a text file with individual genotypes (each row is an individual) to a table of pairwise Fst values. The starting file should NOT have an individual code column (for some reason this gives an error when you turn it into an object of type "loci"). It should just have one column with the genotype (e.g., A/a) and one column with the population. This code was developed for a starting file with a single locus and has not been tested with multilocus data.